examples/sql_plugin/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    /Library/Developer/CommandLineTools/usr/bin/make requirements.txt
#
arrow==1.2.2
    # via jinja2-time
binaryornot==0.4.4
    # via cookiecutter
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via cryptography
chardet==4.0.0
    # via binaryornot
charset-normalizer==2.0.12
    # via requests
checksumdir==1.2.0
    # via nebulakit
click==8.1.2
    # via
    #   cookiecutter
    #   nebulakit
cloudpickle==2.0.0
    # via nebulakit
cookiecutter==2.1.1
    # via nebulakit
croniter==1.3.4
    # via nebulakit
cryptography==37.0.3
    # via secretstorage
cycler==0.11.0
    # via matplotlib
dataclasses-json==0.5.7
    # via nebulakit
decorator==5.1.1
    # via retry
deprecated==1.2.13
    # via nebulakit
diskcache==5.4.0
    # via nebulakit
docker-image-py==0.1.12
    # via nebulakit
docstring-parser==0.13
    # via nebulakit
nebulaidl==0.24.21
    # via nebulakit
nebulakit==0.32.3
    # via
    #   -r ../../../common/requirements-common.in
    #   nebulakitplugins-sqlalchemy
nebulakitplugins-sqlalchemy==0.32.3
    # via -r requirements.in
fonttools==4.32.0
    # via matplotlib
googleapis-common-protos==1.56.0
    # via
    #   nebulaidl
    #   grpcio-status
greenlet==1.1.2
    # via sqlalchemy
grpcio==1.44.0
    # via
    #   nebulakit
    #   grpcio-status
grpcio-status==1.44.0
    # via nebulakit
idna==3.3
    # via requests
importlib-metadata==4.11.3
    # via keyring
jeepney==0.8.0
    # via
    #   keyring
    #   secretstorage
jinja2==3.1.1
    # via
    #   cookiecutter
    #   jinja2-time
jinja2-time==0.2.0
    # via cookiecutter
keyring==23.5.0
    # via nebulakit
kiwisolver==1.4.2
    # via matplotlib
markupsafe==2.1.1
    # via jinja2
marshmallow==3.15.0
    # via
    #   dataclasses-json
    #   marshmallow-enum
    #   marshmallow-jsonschema
marshmallow-enum==1.5.1
    # via dataclasses-json
marshmallow-jsonschema==0.13.0
    # via nebulakit
matplotlib==3.5.1
    # via -r ../../../common/requirements-common.in
mypy-extensions==0.4.3
    # via typing-inspect
natsort==8.1.0
    # via nebulakit
numpy==1.22.3
    # via
    #   matplotlib
    #   pandas
    #   pyarrow
packaging==21.3
    # via
    #   marshmallow
    #   matplotlib
pandas==1.4.2
    # via nebulakit
pillow==9.1.0
    # via matplotlib
protobuf==3.20.0
    # via
    #   nebulaidl
    #   nebulakit
    #   googleapis-common-protos
    #   grpcio-status
    #   protoc-gen-swagger
protoc-gen-swagger==0.1.0
    # via nebulaidl
psycopg2-binary==2.9.3
    # via -r requirements.in
py==1.11.0
    # via retry
pyarrow==6.0.1
    # via nebulakit
pycparser==2.21
    # via cffi
pyparsing==3.0.8
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via
    #   arrow
    #   croniter
    #   nebulakit
    #   matplotlib
    #   pandas
python-json-logger==2.0.2
    # via nebulakit
python-slugify==6.1.1
    # via cookiecutter
pytimeparse==1.1.8
    # via nebulakit
pytz==2022.1
    # via
    #   nebulakit
    #   pandas
pyyaml==6.0
    # via
    #   cookiecutter
    #   nebulakit
regex==2022.3.15
    # via docker-image-py
requests==2.27.1
    # via
    #   cookiecutter
    #   nebulakit
    #   responses
responses==0.20.0
    # via nebulakit
retry==0.9.2
    # via nebulakit
secretstorage==3.3.2
    # via keyring
six==1.16.0
    # via
    #   grpcio
    #   python-dateutil
sortedcontainers==2.4.0
    # via nebulakit
sqlalchemy==1.4.35
    # via nebulakitplugins-sqlalchemy
statsd==3.3.0
    # via nebulakit
text-unidecode==1.3
    # via python-slugify
typing-extensions==4.1.1
    # via
    #   nebulakit
    #   typing-inspect
typing-inspect==0.7.1
    # via dataclasses-json
urllib3==1.26.9
    # via
    #   nebulakit
    #   requests
    #   responses
wheel==0.37.1
    # via
    #   -r ../../../common/requirements-common.in
    #   nebulakit
wrapt==1.14.0
    # via
    #   deprecated
    #   nebulakit
zipp==3.8.0
    # via importlib-metadata
